allow for non-current gmachine to exist
The code assumes that if a gmachine_id exists in the collection, there
exists exactly one record that has 'current:True'. In reality, it can
happen (due to unexpected circumstances). This change allows for this,
by simply skipping the non-current gmachines, instead of raising an
exception. There is now a test for this case as well.

Also added tests for duplicate id in the gid collection