gidmap fixes for slave
when slave, we can't issue a getMore, so we use an aggregate to get a
single record.

Also, sometimes the aggregation would come up empty, or potentially
throw an error. In this case, continue. It might be syncing at the
moment, and the data will be available later.

When slave, treat a missing gid error as non-fatal, so the file will not
end up in the audit, rather tried again next time (hopefully the
collection will sync by then).